pep_regecol

List all endemic species within an ecological region.

List of ecological region ids and full name, based on Zamora, C. 1996.

AA = Altoandina

BHA = Bosques Húmedos Amazónicos

BHMP = Bosques Muy Húmedos Premontanos

BMHM = Bosques Muy Húmedos Montanos

BMHP = Bosques Muy Húmedos Premontanos

BPM = Bosques Pluviales Montanos

BS = Bosques Secos

DCT = Desierto Cálido Tropical

DST = Desierto Semicálido Tropical

MA = Mesoandina

MDE = Matorral Desértico

NDE = Matorral Desértico

PAR = Páramo

PD = Puna Desértica

PSH = Puna Húmeda y Seca

pep_regecol(region = NULL, specie)

Arguments

region

Atomic element or vector with ecological region names, shrot or long name.

specie

Atomic element or vector with species names.

Value

A tibble

References

León, B., J. Roque, C. Ulloa Ulloa, N. C. A. Pitman, P. M. Jørgensen & A. Cano Echevarría. 2006 (2007). El Libro Rojo de las Plantas Endémicas del Perú. Revista Peruana Biol. 13(núm. 2 especial): 1s–971s

ZAMORA, C. 1996. Las regiones ecológicas del Perú. En: Rodríguez, L.O. (ed.), Diversidad Biológica del Perú: Zonas Prioritarias para su Conservación, pp. 137-141. FANPE, GTZ-INRENA. Lima, Perú.

Examples

# By Region ID
pep_regecol("BHA")
#> Error in loadNamespace(x): there is no package called ‘ppendemic’
pep_regecol(c("MA", "BS"))
#> Error in loadNamespace(x): there is no package called ‘ppendemic’

# By species name
pep_regecol(specie = "Aphelandra weberbaueri")
#> Error in loadNamespace(x): there is no package called ‘ppendemic’

especie <-  c("Aphelandra weberbaueri", "Odontophyllum cuscoensis")
pep_regecol(specie = especie)
#> Error in loadNamespace(x): there is no package called ‘ppendemic’